fix: use search param for plugin detail view to avoid @ encoding issues
- Removed broken [plugin]/page.tsx dynamic route (Next.js can't handle @ in
  URL path segments reliably)
- Detail view is now an inline component on the plugins page, driven by
  ?detail=<encoded-id> search param
- No second API fetch needed — data is already loaded by usePlugins()
- Uninstall navigates back to list on success
